requirements.txt
decorator==4.0.10
defusedxml==0.4.1
diff-match-patch==20121119
Django==1.8.2
django-import-export==0.4.5
django-lazysignup==1.0.2
django-modeltranslation==0.10.2
django-select-multiple-field==0.4.2
djangorestframework==3.4.3
flocs==0.0.4
ipython==5.0.0
ipython-genutils==0.1.0
oauthlib==1.1.2
pexpect==4.2.0
pickleshare==0.7.3
pkg-resources==0.0.0
prompt-toolkit==1.0.5
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-social-auth==0.2.19
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
tablib==0.11.2
traitlets==4.2.2
wcwidth==0.1.7
